WIP, fix repeating articles

+++ src/store/mutations.js
@@ -16,10 +16,21 @@ export const mutations = {
             content: payload.extract,
             isChosen: false
         }
-        state.randomArticles.push(articleCard)
-        sessionStorage.setItem('randomArticlesArray', JSON.stringify(state.randomArticles))
-        console.log('saving articles to sessionStorage')
-        console.log(state.randomArticles)
+        // console.log(articleCard)
+        const exists = (element) => {
+            return element.title == articleCard.title
+        }
+
+
+        if (state.randomArticles.find(exists))
+            console.log('it is here')
+        else {
+            state.randomArticles.push(articleCard)
+            sessionStorage.setItem('randomArticlesArray', JSON.stringify(state.randomArticles))
+            console.log('saving articles to sessionStorage')
+            console.log(state.randomArticles)
+        }
+
     },
     readArticleContent(state) {
         if (sessionStorage.getItem('randomArticlesArray'))
